requirements.txt
Python 2.7.10 

Django==1.9.6

djangorestframework==3.5.3


